+++ site/pages/components/Upload/cn.md
@@ -33,6 +33,7 @@
 | multiple | bool | false | 是否支持多选文件 |
 | renderContent | func(res, value, index, values) | - | 自定义结果的内容 |
 | validatorHandle | bool \| func(error, file) | false | 是否处理校验失败的情况, 如果提供一个函数, 则以函数的返回值判断是否处理此 error |
+| dropValidate | func | 无 | 拖拽文件上传前校验，返回 true/false |
 
 
 ### options
